python/frameworks/strands/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "traceai-strands"
version = "0.1.0"
description = "TraceAI instrumentation for AWS Strands Agents"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "FutureAGI", email = "support@futureagi.com" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Monitoring",
]
keywords = [
    "traceai",
    "strands",
    "aws",
    "agents",
    "observability",
    "opentelemetry",
    "tracing",
    "llm",
    "ai",
]

dependencies = [
    "fi-instrumentation-otel>=0.1.11",
    "opentelemetry-api>=1.0.0",
    "opentelemetry-sdk>=1.0.0",
    "opentelemetry-exporter-otlp>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
strands = [
    "strands-agents>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/futureagi/traceai"
Documentation = "https://docs.futureagi.com/traceai"
Repository = "https://github.com/futureagi/traceai"

[tool.hatch.build.targets.wheel]
packages = ["traceai_strands"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
